Fix packet lost after switching by adding a link up delay.
The ethernet link seems to be unstable right after link setup, perhaps
because its partner was not ready when local PHY reports link up.
That is why data sent during this period (about 20ms) might get lost.
As a workaround, this patch add a 100ms delay after link goes up. And it
works.